Refactor test helper for DotNuCompletion
