+++ crates/nu-cli/tests/completions/support/mod.rs
@@ -1,5 +1,3 @@
 pub mod completions_helpers;
 
-pub use completions_helpers::{
-    file, folder, match_dir_content_for_dotnu, match_suggestions, merge_input, new_engine,
-};
+pub use completions_helpers::{file, folder, match_suggestions, merge_input, new_engine};
